services/analise-python/requirements.txt
kafka-python==2.0.2
spacy==3.7.4
# Modelo pré-treinado em português para o spaCy
https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0.tar.gz
# Bibliotecas para o modelo de sentimento Transformer
transformers==4.41.2